move hash{F,P}Small to their own file
